local-backend-huggingface/requirements.txt
gradio>=6.9.0,<7.0
spaces>=0.28.0
transformers>=4.35.0,<4.50
torch>=2.1.0,<2.5
easyocr>=1.7.0,<1.8
kokoro>=0.9.0
soundfile>=0.12.0
opencv-python-headless>=4.8.0,<4.11
numpy>=1.24.0,<2.0
Pillow>=10.0.0,<11.0